Create Variadic function
Variadic functions are functions that take a variable number of arguments. In C programming, a variadic function will contribute to the flexibility of the program that you are developing.

The declaration of a variadic function starts with the declaration of at least one named variable and uses an ellipsis as the last parameter, e.g.

int printf(const char* format, ...);
In this problem, there are three variadic functions named, sum(), min(), and max() to calculate sums, minima, maxima of a variable number of arguments. The first argument passed to the variadic function is the count of the number of arguments, which is followed by the arguments themselves.